Add files via upload
STL and CAD files for mounting bracket that allows a DSLR or other device with a 1/4-20 threaded mount to be attached to a 75 or 100 mm VESA mount.  The mounting holes in the bracket are sized to accept M4 heat-set threaded inserts (6mm OD).  I have only used this bracket when it has been printed using ABS with a 0.4mm nozzle, 0.2mm layer heights, 5 top and bottom layers, 4 perimeters, and 40% infill.  Deviating from these print settings may produce a part that will not sufficiently support your camera.  This part is provided with NO warranty of any sort nor guarantee of suitability for a particular purpose.  If you print & use this part, you should be sure to perform weight checks prior to mounting a camera.  (The bracket will need to support weight much greater than the weight of the camera, so weight checks should be performed with at least 3x the expected maximum weight of the camera that will be mounted.